Fixes #13 by converting init to return a Stream[JValue].
Removed GetQueue and GetModel and rolled the behaviour into AddSite, which avoids having to be careful with the order of the messages.  AddSite now returns the Model and the Queue in one step.

Also converted the init to return a Stream[JValue] as I realized we could create a Stream with the document and the queue in one step.

Added a method to compute the max clock value for a site.  Tried using collection.max but it blows up with exceptions if you have empty collections, which seems crazy.